handle the case that execution path of script bytecode is not fixed (Part 1)
if there is conditional branch or loop in the script, the execution path
will not be a fixed one. when setup parameter for a vm circuit, user can
configure the number of execution steps. Special opcode 'Nop' and 'Stop'
will be used to fill up the execution steps if the number of the actual
execution steps is less than the configured number.